Refactor enums into separate files
