+++ Bot.Shared/Enums/BillerEnum.cs
@@ -1,0 +1,9 @@
+namespace Bot.Shared.Enums;
+
+public enum BillerEnum
+{
+    DSTV,
+    Electricity,
+    Insurance,
+    Water
+}
