+++ Bot.Shared/Enums/DirectDebitStatus.cs
@@ -1,0 +1,10 @@
+namespace Bot.Shared.Enums;
+
+public enum DirectDebitStatus
+{
+    Pending = 0,
+    Success = 1,
+    Failed = 2,
+    Cancelled = 3,
+    Retrying = 4
+}
