+++ Bot.Shared/Enums/PersonalityEnum.cs
@@ -1,0 +1,9 @@
+namespace Bot.Shared.Enums;
+
+public enum PersonalityEnum
+{
+    Formal,
+    Casual,
+    Fun,
+    Friendly
+}
