+++ Bot.Shared/Enums/RewardTypeEnum.cs
@@ -1,0 +1,12 @@
+namespace Bot.Shared.Enums;
+
+public enum RewardTypeEnum
+{
+    Badge,
+    Discount,
+    Cashback,
+    Streak,
+    RecurringStreak,
+    GoalAchieved,
+    Signup
+}
