+++ Bot.Shared/Enums/TransactionStatus.cs
@@ -1,0 +1,9 @@
+namespace Bot.Shared.Enums;
+
+public enum TransactionStatus
+{
+    Pending = 0,
+    Success = 1,
+    Failed = 2,
+    Cancelled = 3
+}
